requirements.txt
jira==3.6.0
python-dotenv==1.0.1
requests==2.31.0
pytest==8.1.1
pytest-mock==3.14.0
loguru==0.7.2
rich==13.7.1
